Task & Subtask

+++ app/Http/Controllers/SubTasksController.php
@@ -39,11 +39,20 @@ class SubTasksController extends Controller
         return view("layouts/tasks/sub.tasks/create");
     }
 
-    public function store(TasksRequest $request) {
+    public function store(TasksRequest $request, $taskId) {
+        $task = Task::where("id", $taskId)->first();
+        if (!$task) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Task not found'
+            ]);
+        }
         $subTask = [
             "name" => $request->input("sub_task_name"),
+            "task_id" => $task->id,
             "image" => $request->file("images"),
             "description" => $request->input("sub_task_description"),
+            "assign_to" => SubTask::$DEFAULT_ASSIGNEE,
             "attachment" => $request->file("attachment"),
             "due_date" => $request->date("due_date"),
         ];
@@ -52,7 +61,6 @@ class SubTasksController extends Controller
 
         Session::flash('success', "Create Sub Task successfully!");
         return redirect()->route("layouts/tasks/sub.tasks");
-        // return response()->json(['success' => true, "message" => "Task created"]);
     }
 
     public function edit() {
@@ -60,8 +68,13 @@ class SubTasksController extends Controller
     }
 
     public function update(TasksRequest $request, $id) {
-        $subTask = SubTask::findOrFail($id);
-        
+        $subTask = SubTask::findOrFail($id)->first();
+        if (!$subTask) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Subtask not found'
+            ]);
+        }
         $subTask->name = $request->input("sub_task_name");
         $subTask->image = $request->file("image");
         $subTask->description = $request->input("sub_task_description");
@@ -75,8 +88,14 @@ class SubTasksController extends Controller
     }
 
     public function delete($id) {
-        $subTask = Task::findOrFail($id);
-        
+        $subTask = SubTask::findOrFail($id)->first();
+        if (!$subTask) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Subtask not found'
+            ]);
+        }
+
         $subTask->delete();
 
         Session::flash("success", "Successfully delete sub task");
@@ -84,7 +103,13 @@ class SubTasksController extends Controller
     }
 
     public function assign(TasksRequest $request, $id) {
-        $task = Task::findOrFail($id);
+        $subTask = Task::findOrFail($id)->first();
         
+        if (!$subTask) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Subtask not found'
+            ]);
+        }
     }
 }
